moved all tests from utils/test_distances.py to lib/test_distances.py
